Small improvements

+++ src/Simulation/Common/IQuantumProcessor.cs
@@ -436,7 +436,7 @@ namespace Microsoft.Quantum.Simulation.Common
         /// <remarks>
         /// A typical implementation will compare all <paramref name="measurementResults"/> to <paramref name="resultsValues"/> and return the result of this comparison.
         /// </remarks>
-        int StartConditionalStatement(IQArray<Result> measurementResults, IQArray<Result> resultsValues);
+        long StartConditionalStatement(IQArray<Result> measurementResults, IQArray<Result> resultsValues);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
@@ -444,11 +444,11 @@ namespace Microsoft.Quantum.Simulation.Common
         /// </summary>
         /// <param name="measurementResult">The actual result of the measurement of a qubit upon which branching is to be performed.</param>
         /// <param name="resultValue">The expected value of result of the measurement of this qubit.</param>
-        /// <returns> A value representing this conditional statement and encoding the result of condition.</returns>
+        /// <returns> A value representing this conditional statement and encoding the result of the condition. It will be passed through to the other branching related APIs such as RunThenClause.</returns>
         /// <remarks>
         /// A typical implementation will compare <paramref name="measurementResult"/> to <paramref name="resultValue"/> and return the result of this comparison.
         /// </remarks>
-        int StartConditionalStatement(Result measurementResult, Result resultValue);
+        long StartConditionalStatement(Result measurementResult, Result resultValue);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
@@ -459,50 +459,50 @@ namespace Microsoft.Quantum.Simulation.Common
         /// <remarks>
         /// A typical implementation will use <paramref name="statement"/> to return whether condition was evaluated to true.
         /// </remarks>
-        bool RunThenClause(int statement);
+        bool RunThenClause(long statement);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
         /// Called when the "then" statement of a conditional statement has finished executing.
         /// </summary>
-        /// <param name="statement">A value representing this conditional statement and encoding the result of condition.</param>
+        /// <param name="statement">A value representing this conditional statement and encoding the result of the condition. This is the value returned from the StartConditionalStatement.</param>
         /// <returns> If true is returned, the "then" statement will be executed again (without calling RunThenClause), folowed by another call to RepeatThenClause.</returns>
         /// <remarks>
         /// A typical implementation will return false.
         /// </remarks>
-        bool RepeatThenClause(int statement);
+        bool RepeatThenClause(long statement);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
         /// Called when the "else" statement of a conditional statement is about to be executed.
         /// </summary>
-        /// <param name="statement">A value representing this conditional statement and encoding the result of condition.</param>
+        /// <param name="statement">A value representing this conditional statement and encoding the result of the condition. This is the value returned from the StartConditionalStatement.</param>
         /// <returns> If true is returned, the "else" statement will be executed, otherwise it will be skipped and RepeatElseClause will not be called.</returns>
         /// <remarks>
         /// A typical implementation will use <paramref name="statement"/> to return whether condition was evaluated to false.
         /// </remarks>
-        bool RunElseClause(int statement);
+        bool RunElseClause(long statement);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
         /// Called when the "else" statement of a conditional statement has finished executing.
         /// </summary>
-        /// <param name="statement">A value representing this conditional statement and encoding the result of condition.</param>
+        /// <param name="statement">A value representing this conditional statement and encoding the result of the condition. This is the value returned from the StartConditionalStatement.</param>
         /// <returns> If true is returned, the "else" statement will be executed again (without calling RunElseClause), folowed by another call to RepeatElseClause.</returns>
         /// <remarks>
         /// A typical implementation will return false.
         /// </remarks>
-        bool RepeatElseClause(int statement);
+        bool RepeatElseClause(long statement);
 
         /// <summary>
         /// Intended for a limited support of branching upon measurement results on a target machine level.
         /// Called when a conditional statement on measurement results has finished executing.
         /// </summary>
-        /// <param name="statement">A value representing this conditional statement and encoding the result of condition.</param>
+        /// <param name="statement">A value representing this conditional statement and encoding the result of the condition. This is the value returned from the StartConditionalStatement.</param>
         /// <remarks>
         /// A typical implementation will clean up any data structures associated with statement.
         /// </remarks>
-        void EndConditionalStatement(int statement);
+        void EndConditionalStatement(long statement);
 
         /// <summary>
         /// Called when <a href="https://docs.microsoft.com/qsharp/api/qsharp/microsoft.quantum.intrinsic.assert">Microsoft.Quantum.Intrinsic.Assert</a> is called in Q#.
